Update README.md
A useful way to visualise a relationship between two variables is to draw its (graph). If the relationship is specified by an equation, then this (graph) is called the (graph) of the equation.

My name is Julio Albano, I am a student at Open University and this my third year. The subject that I am study is BSc (Honours) Natural Science and some of the modules are S111 - Question in Science, MST124 - Mathematic, MST125 - Mathematics and Physics. I am also, study WxMaxima 16.04.2 and this is the subject that I like the most.
A useful way to visualise a relationship between two variables is to draw its (graph). If the relationship is specified by an equation, then this (graph) is called the (graph) of the equation.

My name is Julio Albano, I am a student at Open University and this my third year. The subject that I am study is BSc (Honours) Natural Science and some of the modules are S111 - Question in Science, MST124 - Mathematics, MST125 - Mathematics and Physics. I am also, study WxMaxima 16.04.2 and this is the subject that I like the most.